refactor(oauth): remove oauthdb.listAuthorizedClients
