+++ packages/fxa-auth-server/lib/oauthdb/index.js
@@ -109,14 +109,5 @@ module.exports = (log, config) => {
         payload: oauthParams,
       });
     },
-
-    async listAuthorizedClients(sessionToken) {
-      const oauthParams = {
-        assertion: await makeAssertionJWT(config, sessionToken),
-      };
-      return callRoute('/authorized-clients', {
-        payload: oauthParams,
-      });
-    },
   };
 };
